# (0.90.7) Remove argument splatting in hydrostatic free surface tendency kernel entry functions