--- a/validation/vertical_mixing_closures/test_single_column_model.jl
+++ b/validation/vertical_mixing_closures/test_single_column_model.jl
@@ -1,53 +1,25 @@
 using Oceananigans
-using Oceananigans.Units
 using Oceananigans.TurbulenceClosures: CATKEVerticalDiffusivity
 using Oceananigans.TimeSteppers: time_step!
 
 grid = RectilinearGrid(size=64, z=(-256, 0), topology=(Flat, Flat, Bounded))
 coriolis = FPlane(f=1e-4)
-
-N² = 1e-6
-Qᵇ = +1e-8
-Qᵘ = -2e-4 #
-
-b_bcs = FieldBoundaryConditions(top = FluxBoundaryCondition(Qᵇ))
-u_bcs = FieldBoundaryConditions(top = FluxBoundaryCondition(Qᵘ))
-
 closure = CATKEVerticalDiffusivity()
 
-model = HydrostaticFreeSurfaceModel(; grid, closure, coriolis,
-                                    tracers = (:b, :e),
-                                    buoyancy = BuoyancyTracer(),
-                                    boundary_conditions = (; b=b_bcs, u=u_bcs))
+boundary_conditions = (b = FieldBoundaryConditions(top = FluxBoundaryCondition(1e-8)),
+                       u = FieldBoundaryConditions(top = FluxBoundaryCondition(-2e-4)))
+
+model = HydrostaticFreeSurfaceModel(; grid, closure, coriolis, boundary_conditions,
+                                    tracers = (:b, :e), buoyancy = BuoyancyTracer())
                                     
-bᵢ(z) = N² * z
+bᵢ(z) = 1e-6 * z
 set!(model, b=bᵢ, e=1e-6)
 
-simulation = Simulation(model, Δt=10minutes, stop_iteration=1000)
-
-closurename = string(nameof(typeof(closure)))
-
-diffusivities = (κᵘ = model.diffusivity_fields.κᵘ,
-                 κᶜ = model.diffusivity_fields.κᶜ)
-
-outputs = merge(model.velocities, model.tracers, diffusivities)
-
-simulation.output_writers[:fields] =
-    JLD2OutputWriter(model, outputs,
-                     #schedule = TimeInterval(10minutes),
-                     schedule = IterationInterval(100),
-                     filename = "windy_convection_" * closurename,
-                     overwrite_existing = true)
-
-progress(sim) = @info string("Iter: ", iteration(sim), " t: ", prettytime(sim))
-simulation.callbacks[:progress] = Callback(progress, IterationInterval(100))
-
-@info "Running a simulation of $model..."
-
-time_step!(model, 10minutes)
+# Compile
+time_step!(model, 600)
 
 @time for n = 1:100
-    time_step!(model, 10minutes)
+    time_step!(model, 600)
 end