This PR removes argument splatting in intermediate functions that are called to compute the hydrostatic free surface tendencies.

Argument splatting was removed in a prior PR (that'd be great if someone can remember), but was reinstated in #3360.

This PR re-removes splatting. It yields a 2x performance gain for a column model:

# `main`

```julia
julia> include("test_single_column_model.jl")
┌ Info: Running a simulation of HydrostaticFreeSurfaceModel{CPU, RectilinearGrid}(time = 0 seconds, iteration = 0)
│ ├── grid: 1×1×64 RectilinearGrid{Float64, Flat, Flat, Bounded} on CPU with 0×0×3 halo
│ ├── timestepper: QuasiAdamsBashforth2TimeStepper
│ ├── tracers: (b, e)
│ ├── closure: CATKEVerticalDiffusivity{VerticallyImplicitTimeDiscretization}
│ ├── buoyancy: BuoyancyTracer with ĝ = NegativeZDirection()
└ └── coriolis: FPlane{Float64}...
  0.398369 seconds (727.30 k allocations: 706.103 MiB, 30.55% gc time)
```

# This PR
```julia
julia> include("test_single_column_model.jl")
┌ Info: Running a simulation of HydrostaticFreeSurfaceModel{CPU, RectilinearGrid}(time = 0 seconds, iteration = 0)
│ ├── grid: 1×1×64 RectilinearGrid{Float64, Flat, Flat, Bounded} on CPU with 0×0×3 halo
│ ├── timestepper: QuasiAdamsBashforth2TimeStepper
│ ├── tracers: (b, e)
│ ├── closure: CATKEVerticalDiffusivity{VerticallyImplicitTimeDiscretization}
│ ├── buoyancy: BuoyancyTracer with ĝ = NegativeZDirection()
└ └── coriolis: FPlane{Float64}...
  0.214935 seconds (258.50 k allocations: 195.374 MiB, 10.57% gc time)
```

It also reduces memory allocation. 

Note that the nonhydrostatic model was not changed (it does not splat).