@glwagner 2024-02-13
@simone-silvestri note that the tridiagonal solver also did not re-introduce splatting, so no change needs to be made there.