@navidcy 2024-02-14
> Argument splatting was removed in a prior PR (that'd be great if someone can remember), but was reinstated in #3360.

Did you mean https://github.com/CliMA/Oceananigans.jl/pull/3026 and https://github.com/CliMA/Oceananigans.jl/issues/2996?