@wsmoses 2024-02-14
as mentioned on slack here's @jlk9 's Enzyme integration test PR https://github.com/CliMA/Oceananigans.jl/pull/3480

We should check that the tuple doesn't also break that (or try to find the correct mutually happy solution like maybe vararg type)