@glwagner 2024-02-15
> as mentioned on slack here's @jlk9 's Enzyme integration test PR #3480
> 
> We should check that the tuple doesn't also break that (or try to find the correct mutually happy solution like maybe vararg type)

@jlk9 , the test on #3480 does not pass right now, is that correct?